samples/lab6-final/flaskapi/requirements.txt
flask==3.0.2
celery==5.3.6
psycopg2-binary==2.9.9